requirements.txt
pip==23.1.2
alembic==1.11.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.2
psycopg2-binary==2.9.6
py==1.11.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
sh==2.0.4
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.3.4
